Make alternative wire generation optional
If a face is broken (eg self-intersection) the then called fix method can degenerate the shape to unwanted results. As fallback a wire of the original shape was automatically inserted to the import result.
But often the degeneration is minor and the result is acceptable. Then the additional wires are not needed.